refactor: knowledge and knowledge detail
