+++ app/(dashboard)/dashboard/knowledge/page.tsx
@@ -4,8 +4,8 @@ import { CategoryListRes } from "@/types/category/res"
 import { KnowledgeListRes } from "@/types/knowledge/res"
 import { authOptions } from "@/lib/auth"
 import { getCurrentUser } from "@/lib/session"
-import { CreateKnowledgeButton } from "@/components/app/knowledge/create-knowledge-button"
-import { KnowledgeItemList } from "@/components/app/knowledge/knowledge-item-list"
+import { CreateKnowledgeButton } from "@/components/app/knowledge/operations"
+import { KnowledgeItemList } from "@/components/app/knowledge/ui"
 import { DashboardHeader } from "@/components/header"
 import { BreadCrumbs } from "@/components/pagers/breadcrumb"
 import { DashboardShell } from "@/components/shell"
