+++ components/app/course/detail-sidebar-course.tsx
@@ -11,8 +11,8 @@ import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs"
 import { CourseKnowledgeSectionContent } from "@/components/app/course/detail/course-knowledge/course-knowledge-section-content"
 import { CourseKnowledgeSectionList } from "@/components/app/course/detail/course-knowledge/course-knowledge-section-list"
 import { EmptyKnowledgeCourse } from "@/components/app/course/detail/course-knowledge/empty-course-knowledge-content"
+import { EmptyContent } from "@/components/app/knowledge/detail"
 
-import { EmptyContent } from "../knowledge/detail-sidebar-empty-content"
 import { CourseSectionContent } from "./detail/content/course-section-content"
 import { CourseSectionQuiz } from "./detail/content/course-section-quiz"
 import { EmptyCourseContentInitial } from "./detail/content/empty-course-content-initial"
