+++ components/app/knowledge/detail/content/operations/delete-section-content-sheet.tsx
