+++ components/app/knowledge/detail/content/ui/knowledge-section-content.tsx
@@ -11,9 +11,11 @@ import {
   ContextMenuTrigger,
 } from "@/components/ui/context-menu"
 import { Sheet, SheetTrigger } from "@/components/ui/sheet"
-import { CreateSectionContentSheet } from "@/components/app/knowledge/detail/content/create-section-content-sheet"
-import { DeleteSectionContentSheet } from "@/components/app/knowledge/detail/content/delete-section-content-sheet"
-import { EditSectionContentSheet } from "@/components/app/knowledge/detail/content/edit-section-content-sheet"
+import {
+  CreateSectionContentSheet,
+  DeleteSectionContentSheet,
+  EditSectionContentSheet,
+} from "@/components/app/knowledge/detail/content/operations"
 
 enum ContentType {
   Video = 1,
