+++ components/app/knowledge/detail/knowledge-detail-sidebar-empty.tsx
@@ -1,3 +1,5 @@
+import React from "react"
+
 import { cn } from "@/lib/utils"
 import { Icons } from "@/components/icons"
