+++ components/app/knowledge/detail/section/operations/delete-section-sheet.tsx
