+++ components/app/knowledge/detail/section/ui/empty-knowledge-section-initial.tsx
@@ -2,8 +2,8 @@
 
 import React from "react"
 
-import { EmptyContent } from "@/components/app/knowledge/detail-sidebar-empty-content"
-import { CreateSectionButton } from "@/components/app/knowledge/detail/section/create-section-button"
+import { EmptyContent } from "@/components/app/knowledge/detail/knowledge-detail-sidebar-empty"
+import { CreateSectionButton } from "@/components/app/knowledge/detail/section/operations/create-section-button"
 
 interface EmptyKnowledgeSectionInitialProps {
   id_knowledge: number
