+++ components/app/knowledge/operations/delete-knowledge-button.tsx
@@ -1,7 +1,6 @@
 import React from "react"
 import { useRouter } from "next/navigation"
 
-import { KnowledgeData } from "@/types/knowledge-res"
 import { KnowledgeListResData } from "@/types/knowledge/res"
 import { Button } from "@/components/ui/button"
 import {
