+++ components/app/knowledge/operations/edit-knowledge-button.tsx
@@ -8,8 +8,8 @@ import { Check, ChevronsUpDown } from "lucide-react"
 import { useForm } from "react-hook-form"
 import { z } from "zod"
 
-import { CategoryResponse } from "@/types/category-res"
 import { CategoryListRes } from "@/types/category/res"
+import { KnowledgeListResData } from "@/types/knowledge/res"
 import { cn } from "@/lib/utils"
 import { Button } from "@/components/ui/button"
 import {
@@ -44,18 +44,14 @@ import {
 } from "@/components/ui/sheet"
 import { Textarea } from "@/components/ui/textarea"
 import { toast } from "@/components/ui/use-toast"
-import { CreateButton } from "@/components/create-button"
+import { statusTypes } from "@/components/app/knowledge/operations/create-knowledge-button"
 import { Icons } from "@/components/icons"
 
-interface CreateKnowledgeButtonProps {
+interface DeleteKnowledgeButtonProps {
+  knowledgeData: KnowledgeListResData
   categoryResponse: CategoryListRes
   token: string | undefined
 }
-
-export const statusTypes = [
-  { value: 1, label: "Public" },
-  { value: 2, label: "Private" },
-]
 
 const formSchema = z.object({
   knowledge_title: z
@@ -93,33 +89,36 @@ const formSchema = z.object({
     }),
 })
 
-export function CreateKnowledgeButton({
+export function EditKnowledgeButton({
+  knowledgeData,
   categoryResponse,
   token,
-}: CreateKnowledgeButtonProps) {
+}: DeleteKnowledgeButtonProps) {
   const router = useRouter()
 
-  const [isLoading, setIsloading] = React.useState<boolean>(false)
+  const [isLoading, setIsLoading] = React.useState<boolean>(false)
 
   const [open, setOpen] = React.useState<boolean>(false)
 
   const form = useForm<z.infer<typeof formSchema>>({
     resolver: zodResolver(formSchema),
     defaultValues: {
-      knowledge_title: "",
-      description: "",
-      image: "",
+      knowledge_title: knowledgeData.knowledge_title,
+      description: knowledgeData.description,
+      status: knowledgeData.status,
+      image: knowledgeData.image,
+      id_category: knowledgeData.id_category,
     },
   })
 
   async function onSubmit(values: z.infer<typeof formSchema>) {
-    setIsloading(true)
+    setIsLoading(true)
 
     try {
       const response = await fetch(
-        `${process.env.NEXT_PUBLIC_BASE_URL}/secure/knowledge`,
+        `${process.env.NEXT_PUBLIC_BASE_URL}/secure/knowledge/${knowledgeData.id_knowledge}`,
         {
-          method: "POST",
+          method: "PUT",
           headers: {
             "Content-Type": "application/json",
             Authorization: `Bearer ${token}`,
@@ -130,41 +129,36 @@ export function CreateKnowledgeButton({
 
       if (response.ok) {
         toast({
-          title: "Pengetahuan berhasil dibuat",
-          description: "Pengetahuan berhasil dibuat",
+          title: "Pengetahuan berhasil diubah",
+          description: "Pengetahuan berhasil diubah",
         })
-
-        router.refresh()
         form.reset()
+        router.refresh()
         setOpen(false)
       } else {
-        throw new Error("Gagal membuat pengetahuan")
+        toast({
+          title: "Pengetahuan gagal diubah",
+          description: "Pengetahuan gagal diubah",
+        })
       }
     } catch (error) {
-      toast({
-        title: "Gagal membuat pengetahuan",
-        description: "Gagal membuat pengetahuan",
-        variant: "destructive",
-      })
+      console.error(error)
     } finally {
-      setIsloading(false)
+      setIsLoading(false)
     }
   }
 
   return (
     <Sheet open={open} onOpenChange={setOpen}>
       <SheetTrigger>
-        <CreateButton
-          className=" transition duration-300 delay-150 ease-in-out hover:-translate-y-1 hover:scale-110"
-          name="Tambah"
-        />
+        <Button variant="secondary" className="mr-2">
+          Edit
+        </Button>
       </SheetTrigger>
       <SheetContent position="right" size="content">
         <SheetHeader>
-          <SheetTitle>Tambah Pengetahuan</SheetTitle>
-          <SheetDescription>
-            Tambah pengetahuan baru untuk dipelajari oleh user.
-          </SheetDescription>
+          <SheetTitle>Ubah Pengetahuan</SheetTitle>
+          <SheetDescription>Ubah pengetahuan yang sudah ada.</SheetDescription>
         </SheetHeader>
         <Form {...form}>
           <form
@@ -183,7 +177,7 @@ export function CreateKnowledgeButton({
                     <Input placeholder="Pendahuluan" {...field} />
                   </FormControl>
                   <FormDescription>
-                    Judul pengetahuan yang ingin dibuat.
+                    Judul pengetahuan yang ingin diubah.
                   </FormDescription>
                   <FormMessage />
                 </FormItem>
@@ -206,7 +200,7 @@ export function CreateKnowledgeButton({
                     />
                   </FormControl>
                   <FormDescription>
-                    Deskripsi pengetahuan yang ingin dibuat.
+                    Deskripsi pengetahuan yang ingin diubah
                   </FormDescription>
                   <FormMessage />
                 </FormItem>
@@ -226,7 +220,7 @@ export function CreateKnowledgeButton({
                     />
                   </FormControl>
                   <FormDescription>
-                    Link gambar pengetahuan yang ingin dibuat.
+                    Link gambar pengetahuan yang ingin diubah.
                   </FormDescription>
                   <FormMessage />
                 </FormItem>
@@ -380,7 +374,7 @@ export function CreateKnowledgeButton({
               {isLoading ? (
                 <Icons.spinner className="h-5 w-5 animate-spin" />
               ) : (
-                "Tambah"
+                "Ubah"
               )}
             </Button>
           </form>
