+++ components/app/public-knowledge/public-knowledge-detail-sidebar.tsx
@@ -14,8 +14,7 @@ import { Button } from "@/components/ui/button"
 import { Card } from "@/components/ui/card"
 import { ScrollArea } from "@/components/ui/scroll-area"
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs"
-
-import { EmptyContent } from "../knowledge/detail-sidebar-empty-content"
+import { EmptyContent } from "@/components/app/knowledge/detail"
 
 export default function PublicDetailSidebarKnowledge(props: {
   dataKnowledge: KnowledgeByIdResponse
